fix deleteTodo & add restoreTodo

+++ src/screens/Lists/Components/ListItem.tsx
@@ -7,7 +7,7 @@ import {RectButton} from 'react-native-gesture-handler';
 import {Icon} from 'react-native-elements';
 import {useDispatch} from 'react-redux';
 import {ITodo} from '../../../store/types/todosTypes';
-import {deleteTodo} from '../../../store/actions/todosActions';
+import {deleteTodo, restoreTodo} from '../../../store/actions/todosActions';
 
 interface ITodoStatus {
   status: 'Not started' | 'In progress' | 'Completed';
@@ -18,6 +18,7 @@ export const ListItem = (props: ITodo) => {
   const dispatch = useDispatch();
   const [createdAt, setCreatedAt] = useState('');
   const [status, setStatus] = useState<ITodoStatus['status']>('Not started');
+  const [isDeleted, setIsDeleted] = useState(false);
 
   const theme = StyleSheet.create({
     row: {
@@ -40,15 +41,11 @@ export const ListItem = (props: ITodo) => {
       color: colors.darkGrey,
     },
     buttonDelete: {
-      backgroundColor: colors.error,
+      backgroundColor: isDeleted ? colors.successMain : colors.error,
     },
     rightButtonsBorder: {
       borderColor: colors.grey,
       borderLeftWidth: 1,
-    },
-    leftButtonsBorder: {
-      borderColor: colors.error,
-      borderRightWidth: 1,
     },
   });
 
@@ -65,25 +62,37 @@ export const ListItem = (props: ITodo) => {
       setCreatedAt(d);
       setStatus(s);
     }
-  }, []);
+  }, [props.started_at, props.finished_at, props.created_at, props.completed]);
+
+  useEffect(() => {
+    setIsDeleted(props.category === 'deleted');
+  }, [props.category]);
 
   const renderLeftActions = (_, dragX: any) => {
     const transition = dragX.interpolate({
-      inputRange: [0, 50, 100],
+      inputRange: [0, 60, 61],
       outputRange: [0, 1, 1],
     });
 
-    const handleDeleteTodo = () => {
-      dispatch(deleteTodo(props.id));
+    const handleAction = () => {
+      if (isDeleted) {
+        dispatch(restoreTodo(props.id));
+      } else {
+        dispatch(deleteTodo(props.id));
+      }
     };
 
     return (
       <View style={styles.inline}>
         <RectButton
           style={[theme.button, theme.buttonDelete]}
-          onPress={handleDeleteTodo}>
+          onPress={handleAction}>
           <Animated.View style={{opacity: transition}}>
-            <Icon type={'material'} name={'delete'} color={colors.white} />
+            <Icon
+              type={'material'}
+              name={isDeleted ? 'restore-from-trash' : 'delete'}
+              color={colors.white}
+            />
           </Animated.View>
         </RectButton>
       </View>
@@ -91,9 +100,22 @@ export const ListItem = (props: ITodo) => {
   };
 
   const renderRightActions = (_, dragX: any) => {
-    const trans = dragX.interpolate({
-      inputRange: [0, 50, 100, 101],
-      outputRange: [-20, 0, 0, 1],
+    // TODO: must depend from screen size
+    const first = dragX.interpolate({
+      inputRange: [-61, -60, -30],
+      outputRange: [1, 1, 0],
+    });
+
+    // TODO: must depend from screen size
+    const second = dragX.interpolate({
+      inputRange: [-131, -130, -100],
+      outputRange: [1, 1, 0],
+    });
+
+    // TODO: must depend from screen size
+    const third = dragX.interpolate({
+      inputRange: [-181, -180, -150],
+      outputRange: [1, 1, 0],
     });
 
     const changeTodoState = () => {
@@ -111,23 +133,23 @@ export const ListItem = (props: ITodo) => {
     return (
       <View style={styles.inline}>
         <RectButton style={theme.button} onPress={completeTodo}>
-          <View style={theme.rightButtonsBorder}>
+          <Animated.View style={[theme.rightButtonsBorder, {opacity: third}]}>
             <Icon type={'material'} name={'done'} color={colors.darkGrey} />
-          </View>
+          </Animated.View>
         </RectButton>
         <RectButton style={theme.button} onPress={editTodo}>
-          <View style={theme.rightButtonsBorder}>
+          <Animated.View style={[theme.rightButtonsBorder, {opacity: second}]}>
             <Icon type={'material'} name={'edit'} color={colors.darkGrey} />
-          </View>
+          </Animated.View>
         </RectButton>
         <RectButton style={theme.button} onPress={changeTodoState}>
-          <View style={theme.rightButtonsBorder}>
+          <Animated.View style={[theme.rightButtonsBorder, {opacity: first}]}>
             <Icon
               type={'material'}
               name={'play-arrow'}
               color={colors.darkGrey}
             />
-          </View>
+          </Animated.View>
         </RectButton>
       </View>
     );
